requirements.txt
requests==2.28.2
dns-lexicon==3.11.7
## The following requirements were added by pip freeze:
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
filelock==3.9.0
idna==3.4
importlib-metadata==6.0.0
pycparser==2.21
PyYAML==6.0
requests-file==1.5.1
six==1.16.0
soupsieve==2.3.2.post1
tldextract==3.4.0
urllib3==1.26.14
zipp==3.11.0
